+++ lib/fetch_util/browser/stabilization.rb
@@ -5,9 +5,11 @@ module FetchUtil
     module Stabilization
       autoload :PageFlow, "fetch_util/browser/stabilization/page_flow"
       autoload :SpaHydration, "fetch_util/browser/stabilization/spa_hydration"
+      autoload :Anubis, "fetch_util/browser/stabilization/anubis"
 
       include PageFlow
       include SpaHydration
+      include Anubis
     end
   end
 end